Big one...
Created shop. it uses the item_pool and so it's rarity system.
This version has returned no compilation errors nor any errors on the tests.

this commit includes :
the modifications to the interface to now support the shop.
some new comments to be more readable
new shop interfaces to the interface_test.c file.
and the shop.c implementation itself.

on the shop you can :
- buy
- sell
- refresh
- restock